refactor(dashboard): streamline support drawer by consolidating constants and components for improved maintainability
